move wrapfor to hiveinspector to reuse them
